Refactored workflow name and fixed tb config path
