+++ cg/meta/workflow/balsamic.py
@@ -8,7 +8,7 @@ from typing import List, Optional
 from cg.constants import DataDelivery, Pipeline
 from cg.exc import BalsamicStartError
 from cg.meta.workflow.analysis import AnalysisAPI
-from cg.meta.workflow.fastq import BalsamicFastqHandler
+from cg.meta.workflow.fastq_handler import BalsamicFastqHandler
 from cg.models.cg_config import CGConfig
 from cg.store import models
 from cg.utils import Process
