+++ tests/cli/workflow/fastq/conftest.py
@@ -4,31 +4,31 @@ import datetime as dt
 
 from cgmodels.cg.constants import Pipeline
 
-from cg.meta.workflow.send_fastq import SendFastqAnalysisAPI
+from cg.meta.workflow.fastq import FastqAnalysisAPI
 from cg.models.cg_config import CGConfig
 from cg.store import models
 from tests.store_helpers import StoreHelpers
 
 
-@pytest.fixture(name="send_fastq_context")
-def fixture_send_fastq_context(
+@pytest.fixture(name="fastq_context")
+def fixture_fastq_context(
     cg_context: CGConfig,
     fastq_case,
     helpers: StoreHelpers,
     case_id,
     tb_api,
 ) -> CGConfig:
-    """Returns a CGConfig where the meta_apis["analysis_api"] is a SendFastqAnalysisAPI and a store containing a
-    fastq case"""
+    """Returns a CGConfig where the meta_apis["analysis_api"] is a FastqAnalysisAPI and a store
+    containing a fastq case"""
     _store = cg_context.status_db
     cg_context.trailblazer_api_ = tb_api
-    send_fastq_analysis = SendFastqAnalysisAPI(config=cg_context, pipeline=Pipeline.FASTQ)
+    fastq_analysis = FastqAnalysisAPI(config=cg_context, pipeline=Pipeline.FASTQ)
 
     # Add fastq case to db
     fastq_case["samples"][0]["sequenced_at"] = dt.datetime.now()
     helpers.ensure_case_from_dict(store=_store, case_info=fastq_case)
 
-    cg_context.meta_apis["analysis_api"] = send_fastq_analysis
+    cg_context.meta_apis["analysis_api"] = fastq_analysis
     return cg_context
