+++ tests/meta/workflow/test_fastq.py
@@ -1,16 +1,16 @@
-from cg.meta.workflow.send_fastq import SendFastqAnalysisAPI
+from cg.meta.workflow.fastq import FastqAnalysisAPI
 from cg.store import models, Store
 from tests.cli.workflow.conftest import tb_api
-from tests.cli.workflow.send_fastq.conftest import fixture_send_fastq_context, fixture_fastq_case
+from tests.cli.workflow.fastq.conftest import fixture_fastq_context, fixture_fastq_case
 
 
 def test_upload_bundle_statusdb(
-    case_id: str, customer_id, mocker, send_fastq_context, ticket_nr: int, tmpdir_factory
+    case_id: str, customer_id, mocker, fastq_context, ticket_nr: int, tmpdir_factory
 ):
-    """Tests the status_db upload for the SendFastqAnalysisAPI"""
-    # GIVEN that a send_fastq_context
-    send_fastq_api: SendFastqAnalysisAPI = send_fastq_context.meta_apis["analysis_api"]
-    status_db: Store = send_fastq_context.status_db
+    """Tests the status_db upload for the FastqAnalysisAPI"""
+    # GIVEN that a fastq_context
+    fastq_api: FastqAnalysisAPI = fastq_context.meta_apis["analysis_api"]
+    status_db: Store = fastq_context.status_db
 
     # GIVEN a case with no analyses
     for analysis in status_db.family(internal_id=case_id).analyses:
@@ -18,12 +18,12 @@ def test_upload_bundle_statusdb(
     ticket_dir: str = tmpdir_factory.mktemp(str(ticket_nr))
 
     # GIVEN that a delivery folder is present
-    mocker.patch.object(SendFastqAnalysisAPI, "get_deliverables_file_path")
-    SendFastqAnalysisAPI.get_deliverables_file_path.return_value = ticket_dir
+    mocker.patch.object(FastqAnalysisAPI, "get_deliverables_file_path")
+    FastqAnalysisAPI.get_deliverables_file_path.return_value = ticket_dir
     assert status_db.family(internal_id=case_id).analyses == []
 
     # WHEN the upload_bundle_statusdb is called
-    send_fastq_api.upload_bundle_statusdb(case_id=case_id)
+    fastq_api.upload_bundle_statusdb(case_id=case_id)
     case_obj: models.Family = status_db.family(internal_id=case_id)
 
     # THEN the analysis should be in statusdb and have both the upload_started at and the uploaded_at fields filled.
